[Refactor] frontend/streamlit_prototype/service에 있던 모든 파일을 components로 이동
